My goal here was to make the PHC main page more useful and friendly for people looking for documentation of a specific area.  A user survey/documentation/discussion/proposal type of process would probably have been a good idea, but given that I'd just completed the Plone Boot Camp and had a prime chance to use some of the stuff I'd just learned, I couldn't pass up a chance to actually work in some real code.  I hope but do not expect that some of the work here might actually be interesting.
The one major change I made is simple: list each help type's sections
on the main help center view, and make these sections working links.
This had unexpected wrinkles (i18n translation stuff, in-page link
oddities) but wasn't that bad to do.





helpcenter_view.pt

There are more changes that should probably be considered for this
page, among them a bit of text at the top listing the help types (FAQ,
tutorial, etc.) and maybe descriptions of the help types at the
bottom.  I didn't add many words because I ran into the i18n thing
and, since I don't know Italian or Chinese well enough to do the
translations myself, thought better of barging on ahead with adding
text only in English.  Although I did add one or two things and change
one text string (as you can see in my i18n notes).



What did change:

Mainly, each help type now gives a list of the active sections in that help
type with the number of items in each section.  If there is only one
section, then the individual section name is not given, but the number
of items is still listed.

I also moved the RSS listing and help link to below the main help type
areas; the help link is mainly "how to contribute"

I went ahead and removed the most recent listings (saved this format
by copying to a new view name for easy reference, but it's not
integrated in anywhere).  It's easy to add the top three or five
listings back in below the section listings if that is desired.

I'm not happy with most of the HTML/CSS presentation here or in the
help type folders; I'm not familiar enough with the CSS scheme here to
choose something good.  In particular, the section list has too much
"leading", I think, since section names may be shorter than the full
document titles that were in their place.  OTOH, my sample data may
just not be long enough.  Again, I'd be glad to take a suggestion and
quickly implement it here.

There are a pair of tal:condition statements in there that are
essentially an if/then/else structure.  Perhaps I should implement
this another way since tal:condition isn't intended to be used this
way - would appreciate feedback on this.  It's basically this:
If the glossary (for example) has more than one section, then the
sections are listed with the number of items in each section; else,
the total number of items for the help type is given (no sections are
listed)

Glossary page is mysterious to me.  When one looks at it in a browser,
each glossary heading is followed by a space and then '#'.  See
http://plone.org/documentation/glossary
There's a link around this # - an <a> tag that contains both a name
and an href.  I can guess that this could be useful, but why put it
around the #?  The href is set to exactly the location that the name
gives - the tag links to itself.  Why not just put the name in the <a>
tag around the glossary entry title?  Here's an example of the output,
from the top of the Plone documentation page:

	<a href="http://plone.org/documentation/glossary/zodb">ZODB</a> <a
	class="link-anchor"
	href="http://plone.org/documentation/glossary#zodb" name="zodb">#</a>




The six main help type folder views:

howtofolder_view.pt
faqfolder_view.pt
glossary_view.pt
tutorialfolder_view.pt
helplinkfolder_view.pt
errorreferencefolder_view.pt

The code I added/changed in each of these should be abstracted out
into some common thing, but these six templates are seem so similar to
each
other that I wonder if there are subtle differences that would make
such abstraction not straightforward.  With more time, I'd pick
through them and see if I could do it the right way, but for tonight I
have to leave them as-is (this was a learning exercise for me) with
common code that I pasted in, identical in each view (I hope and
believe).

The code does this: puts a list of section links at the top of each
listing, with each link going to the actual section below.  Each of
the sections in the body of the page now has an anchor just above it
(with a space in the anchor) to be the target of the link.

This list of links is currently in <ul><li></ul> format; I could
probably come up with something better given time, but I suspect that
someone like limi could do so in an instant.  I'd be glad to take
a suggestion and quickly implement it.

The in-page link problem, where the link-constructing code doesn't
know if the page URI ends in a / or not, is only a problem when the
page URI _does_ end in a slash.  This happens when the page is called
explicitly with a slash at the end (e.g. myphc/faq/), or after the
main folder view (e.g. the Faq folder) is edited - after the edit is
completed, the URI will be sufficiently complex that the page links
don't realize they're going to the same page.  As geoffd pointed out,
the worst that happens is that a new page is loaded; while this isn't
deadly, it is highly annoying.  The links are constructed using the
current_page_url variable, so if that somehow adapts, the problem may
be magically fixed.




i18n changes

phc-en.po - Added some things to the i18n english file so that the
main help
center page could label its sections "faqs", "glossaries", etc. - some
of these were present already, some not.  The auto-generated id for
labeling the links section already existed, but wasn't what is wanted
-
needed to address this, because the section was labeled "link:"
(instead
of "links") on the PHC main page.  That phc_link item was used only in
helplink_view.pt, that I could find, so I changed that label name to
phc_link_indicator (it was a label before the actual link in the
listing) and made phc_link be what I wanted.

--> must do this in the other i18n files!  I could do it, but would
want to not be too tired when messing with other people's languages.
Also, it didn't seem worth it until/unless the changes I made are
actually interesting to others.

Added i18n value 'phc_items' to be used for the word 'items' on the
main page.



types/PHCFolder.py

I don't believe I changed this file; I may have accidentally touched
it while looking around for a function to call.