Fixed download templates so that they work with the accessor and not the attribute of downloadableFile, since this was tripping up external storage.
Also changed functionality slightly: it now defaults to internal file if
both external and internal are specified. Since we are wasting our server
space on these, it seems sensible to prioritise them over external links.

Added workaround for the fact that primary=1 doesn't seem to work on the
file field. Any light to be shed on this would be nice. Don't know if it's
external-storage related or not.