* Fixed some minor issues with the relative url and CMF uid handling in    ATFavorite.    [tiran]
 * Added AT and CMF uid migration to the migration system.
   [tiran]

 * Fixed the unit test fixture and ATFavorite unit tests. Now tests aren't
   running as Manager only when really needed.
   [tiran]

I beleave that the CMFUid tools are someward strange to broken. Sometimes all unit tests are passing and sometimes 1 or 2 favorite unit tests are failing. Either all unit test have to pass or two favorite tests may fail. The CMFUid handling seems to be timing critical which smells like a threading issue but I can't find the reason for it.