Removed the silent option on rebuilt, merge and sync completly. The addition of the added- and removed-sections lead to an incorrect format.
PYReader now extracts the line numbers of messages and writes these at the end of the reference seperated with a ':'

The catalog's add method doesn't adds duplicate references or automatic comments anymore.

Fixed handling of normal comments in PT- and PYReader.

Always add a blank line as the last line on po's, as poEdit does it

Adjusted Python parsing tests to cover the current behaviour. Added XXX comments where it is wrong

Fixed an issue with an unnecessary whitespace in messages [thx Tuttle for the test]