First cut of a reference browser using Kupu drawers. Right now it isn't terribly friendly (it nukes ATReferenceBrowser), but that will change.
Drawers now properly support multiple selections. Also any selected references are checked on entry to the appropriate library or folder.

Bookmark button in linkdrawer is now only shown if the drawer supports bookmarks (links do, references don't), and if the xml includes a <bookmark/> tag for that resource. kupuInfoForBrains.py only sets bookmark if the object type is in the 'bookmarkable' collection.