Fix for an issue that tomster found.  Shouldn't have been calling Weblog.getAllEntries from the template as this returns unpublished entries as well.  Unpublished entries have no 'effective' date, and consequently get screwy with code that expects published items.
The fix was to use Weblog.getEntries instead, however, I have changed the behaviour of this method.  It now returns a simple sequence of entries (reasonably enough), as opposed to a dictionary of date-keyed entries.  The latter functionality was moved to a utility method on the quills tool by an earlier checkin, so templates wanting this form of sorting just need to pass in the result of a Weblog.getEntries call to QuillsTool.sortWeblogEntriesByDate.

I didn't find any other place in Quills where Weblog.getEntries is called, so this (hopefully) won't affect anyone.