Adding some notes about how view support may work. The idea is that IBrowserDefault and the FTI should support vocabularies with @@view_name, and look up the view if that's the case.
There are some considerations:

 - CMF method aliases: Ideally, it should be possible for these to point to
   views. This should be fixed in CMF, though, if it doesn't work already.
 - View titles: getAvailableLayouts() needs to return titles, which means
   we need a way of getting a human-readable, i18n aware title
 - __call__() should look up views and call them when appropriate
 - Testability - this code has inadequate tests! Doctests may be easier than
   unit tests.

Ultimately, the use in Plone depends on CMF method aliase. The "display" menu
simply does get/setLayout() and uses getAvailableLayouts() as the query. When
viewed, the (Default) alias points to (dynamic view); this delegates to
PloneTool.browserDefault(), which calls getLayout() if there is no default page.
getLayout() in this case would just return the @@view_name. That is basically
equivalent to calling /path/to/object/@@view_name, so in theory all that
*ought* to work, but we have to test it. :)

If it *does* work, the only challenges are (a) view titles (b) supporting 
__call__ (which is a fallback and should be easy enough via zapi.queryView()),
and (c) writing some real tests :-)

(who said a checkin message couldn't be an essay?)