Use text/x-html-safe as the output type for the 'text' field.
This makes good sense in and of itself, but also (due to the slightly annoying deficiency of AT/PortalTransforms) makes it easier to hook in other transforms (with PT 'policies') for WeblogEntry output.  In my case, I'd like to be able to cite CMFBibliographyAT references from my blog entries, and this change makes it possible to hook in the CMFBib transform that does this.

Also, removed a backwards compatible method alias because I just don't think it's necessary.  Speak up if you disagree.