Adding a rudimentary migration framework between versions of CMFBib.  Migrations now live in CMFBibliographyAT/migrations/.  They *must* be safe to call multiple times!
Use this migration framework to add a _reference_types attribute to the portal_bibliography tool in order to keep track of portal_types that should be treated as references.  This makes it possible to make a portal_type addable to BibFolders without them needing to be actual reference-ish objects.

Also added a separate permissions.py module to keep permissions that the project uses in a separate place.  This should probably mutate into permissions.zcml (or similar) in the future.