Bring some sanity to the workflows and permissions. A few notes:
o Reverting to CMF core permissions as much as possible. Everything else is just
not worth it. 

o Having AddComment be the write_permission of the text field was a hack :)

o There seems to be some bug in _createObjectByType() or Archetypes which means
that the permissions are always acquired from the parent when you try to mutate
fields (via the call to update()). Thus, we need to call mutators directly and
not pass kwargs to _createObjectByType(). At least that's the best explanation
I could think of :)

o There seems to be some race condition in the code that affects the ordering
of comments coming out of getComments(). Basically, 'created' only goes down
to the hour and minute (?) and so the ordering is wrong in some cases. This
is affecting the ftests so you may get random failures :-(