(1) Translatable Bibliography: added two boolean switches to the bibliography tool that can be used to enable/disable LinguaPlone support.
(2)
BibFolders/BibrefItems: added an isTranslatable() method that overrides
PLONE's default isTranslatable.py script. It additionally queries the
translatability configuration of the bibliography setup.

(3)
prefs_bibliography_memberrefsupport_form.cpt: Added the respective
radio buttons that configure translatability to the plone configlet
(Bibliography Setup - Details)

(4)
inferAuthorReferences: complete rewrite of the code, takes several
information into account (member type fields, authors field, Title()), 
before it gives up.

(5)
inferAuthorReferences: also sets the creators field (if available),
added a switch AUTHOROF_IMPLIES_CREATOR to config.py

(6)
member reference support: made the mechanism LinguaPlone robust.
If the referred to member type is a multi-lingual object, we
now always refer to the canonical object. We also derive the author
data from the canonical object.

(7)
The reference that is stored in the author schema is also a
reference to the member type's canonical object.
The getSiteMember method also generates an author list (for the
selector) that goes back to contents of the resp. the canonical
objects.

The necessary changes occur in BaseEntry.getSiteMembers() and
BaseEntry.pre_validate() and BaseEntry.inferAuthorReferences

(8)
getMembers: commented out the getMembers method, as it does not really
seem to be used, seems to be working fine without...

(9)
parsers/renderers: added the proposed isAvailable() method to
parsers and renderers. isAvailable() is always True in the base code
of parsers and renderers, but can be overriden in the parsers' /
renderers' specific code snippets. Target use of isAvailable() is a
pre-check, if all requirements for usage of a specific parser /
renderer are fullfilled (e.g. AmazonTool installed and configured,
TODO: LaTeX installed, etc.).