Adding a "DOI" StringField to ArticleReference.  Looking at http://www.doi.org/, there's probably a case for adding this field to (all?) other types as well, but I've only ever had cause to use it for articles, so I'll just stick with that for now.
There are no tests with this checkin as it solely adds a field.  This is an extension of the implicit interface for ArticleReference.

No support for renderers or parsers of any sort was added here, so CMFBib remains clueless about DOIs outsite of Plone.  That's a shame, but I don't have the spare time to look into that.

Finally, as DOIs are effectively UUIDs, I would have thought they have some potential for use in the duplicate detection code.  I'm afraid I have to leave that as an exercise to the reader as well :(.