The testObjectImplements test tries to adapt all the interfaces indexed for the membrane_tool.  But the registrations for IContentMenuItem in plone 3.0 can't be adapted.
We can't just reduce the index to include only directly implemented
interfaces as Rob tells me that there is code that still requires the
index to work for interfaces for which multi-adapters are registered
that the object can satisfy itself.

We also cannot determine from the registrations themselves what the
object can and can't be adapted to, so we have to query the adapters.

This is probably unacceptably inefficient but Rob tells me that the
future will be to reduce the requirements to just directly implemented
interfaces so this will do for now.