Added a SSHADigestEncrypter that works just like Zope's SSHADigestScheme. This makes it easy for people to migrate ZODB users to SQL users and keep on using the same password scheme.
I had to extend the encrypter interface to also do the validation.
The reason for this is that the consumer of the encrypter interface
should not have to know how to validate something it did not encrypt in
the first place. Given this functionality the interface name is less
accurate but renaming has more implications than is necessary to deal
with at once.