Implementing require_login in SessionCrumbler, to fix certain errors that made PAS traverse to credentials_basic_auth and never displaying the insufficient privileges page. require_login traverses as nescessary, and locks in to credentials_session_auth in a better way.
There's more work needed, but all this is done for Plone3. I just thought some better behavior for 2.5.* was nice. It should really implement a bunch of 
interfaces too, but it's all done in 3.

This breaks Plone 2.0, so there might be a need for a fix, or at least don't set the next release as working with it :)