Set PageBuilder-objects as non-structural folderish objects.
Added custom add-menu to allow creating objects within the PageBuilder-
object from the document content menu. Currently, this is implemented
using javascript; we create the add-menu and move it dynamically into
the content menu after the page is loaded.

It is neccesary to reinitialize the dropdown-menus to enable the
correct handlers.