Fix a bug with GS step order
If archetypetool.xml registers a type for the catalog map and
membranetool.xml is run after archetypetool.xml, then membranetool.xml
still needs to register the types with the status map.

This problem only surfaces when the archetypetool step is run before
the membranetool step and since the order is unpredictable this test
exposes the bug.