Refactor intid into it's own package.
Done in preparation for supporting intid userids.

Note that the use of ZCML overrides means that total separation is not
possible or at least not easy.  ZCML overrides must be collected at
the time that all site ZCML is loaded and will not work in a file
loaded in a test with Products.Five.zcml.load_config().  Since we have
to include overrides in the root sharedusers *product* ZCML, we also
have to include the rest of the intid stuff in sharedusers root
including the GS componentregistry utility.