Fix a serious packer bug with full packing.
Any variable names starting with _ are transferred their names with
a symbol table. This is usually good, except when not: for example
we want to read a variable from dom, whose name starts with _.

More surprisingly, the variable names are mangled out of
strings too - a bug in the current packer or we need to update it.