Removing MetaWeblogAPI.py from Quills.  The functionality is now provided by a view found in quills.remoteblogging.
Add MetaWeblogPASPlugin as a hard dependency for now.  This should be toned down to a soft dependency in the not-too-distant future.

Add a migration step to remove the old 'metaWeblog' and 'blogger' attributes on Weblog instances.  Also add a test for this migration step.

Remove MetaWeblogAPI stuff from Install.py.  Also make the config imports a little clearer.

Also, disable the 0.9->1.5 migration step, for now.  The code in there has expectations about code and attributes being available that just aren't any more.  We need to rethink the migration stuff a little.