Add the 'avoidPhaseOneInit hack to the plone-initKupu action
this is meant to disable double initialization of kupu
in phase 1, in older kupu varsions that do not contain the code
guarding a double init.

This would fix a bug when for example the Bold and Italic
buttons seem to do nothing where in reality they are triggered
twice at the same time and preserving the original text style.