Refactor metaweblogapi implementation.  This moves us closer to having a generic implementation that only relies on the IWeblog and IWeblogEntry interfaces, rather than the Products.Quills implementation-specific code that went before.
It's not finished yet as I still need to figure out some way of handling the 'publish' step.  This raises a broader question of how to deal with workflow more generally.  Currently, the IWeblog interface knows about workflow in the sense that it has a 'getDrafts' method.  Following that pattern, then IWeblogEntry should grow a 'publish' method.  However, fschulze doesn't like this mixing of workflow with more fundamental weblog interfaces.  I recognise the concern, but then I think the notion of drafts and publishing is fairly basic to blogging activities, and as such may be a valid part of the core weblog interfaces.  Opinions most welcome!