Previous fix didn't work.  Fixed now by sorting the groups.
Unfortunately, I can't figure a way to test for this bug because the
catalog results are unsortable.  I'd have to do way too much
scaffolding to test for this.  See
Products/gpcore/tests/addallbcs_directory.txt for a test that
reproduced it at least a few times for a few people.