Moving the functional tests out of the way for the time being.  We aren't currently running the z3 tests, and need to be able to see unit test failures in this module.
As noted at the top of the file:

""" NB - this file has been renamed in order to prevent the enclosed
tests from running (and failing).  At this date, we are not testing
in Zope 3 - which these tests are intended to exercise.  Bring back
to life when ready by renaming this file to:
  testFunctional.py
"""