The local role migrator presumptuously assigns an 'Owner' local role if none is already assigned.  This results in a local role being assigned for None if the owner can no longer be found which causes errors in the allowedRolesAndUsers index.
The Owner local role should probably not be assigned in the first
place since it is legal for content to have no local roles.  Since
there may be code in the wild depending on this, however, this commit
simply adds a check for missing owner and sets no local roles in that
case.